requirements.txt
ultralytics>=8.3.0
streamlit>=1.43.0
Pillow>=10.0.0
PyYAML>=6.0.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.8.0
